+++ pytket/binders/circuit/main.cpp
@@ -500,6 +500,10 @@ PYBIND11_MODULE(circuit, m) {
           "StatePreparationBox", OpType::StatePreparationBox,
           "A box for preparing quantum states using multiplexed-Ry and "
           "multiplexed-Rz gates")
+      .value(
+          "DiagonalBox", OpType::DiagonalBox,
+          "A box for synthesising a diagonal unitary matrix into a sequence of "
+          "multiplexed-Rz gates")
       .def_static(
           "from_name", [](const json &j) { return j.get<OpType>(); },
           "Construct from name");