+++ tket/src/Transformations/include/Transformations/Decomposition.hpp
@@ -245,8 +245,7 @@ Transform decomp_arbitrary_controlled_gates();
 // and adjust the direction of their decomposition (i.e. CnX = CnX.dagger)
 // to improve the chance of gate cancellation. This method will not improve
 // the decomposition when the CnX gates are scattered; but it works the best
-// when CnX gates are very close to each other, such as the circuit produced by
-// a ToffoliBox.
+// when CnX gates are very close to each other.
 Transform cnx_pairwise_decomposition();
 
 }  // namespace Transforms