+++ tket/src/Utils/HelperFunctions.cpp
@@ -55,4 +55,15 @@ std::vector<bool> dec_to_bin(unsigned dec, unsigned width) {
   return bits;
 }
 
+// big endian
+unsigned bin_to_dec(const std::vector<bool> &bin) {
+  unsigned res = 0;
+  for (unsigned i = 0; i < bin.size(); i++) {
+    if (bin[i]) {
+      res = res + (1 << (bin.size() - 1 - i));
+    }
+  }
+  return res;
+}
+
 }  // namespace tket