+++ tket/src/Utils/include/Utils/HelperFunctions.hpp
@@ -88,4 +88,10 @@ uint32_t reverse_bits(uint32_t v, unsigned w);
  */
 std::vector<bool> dec_to_bin(unsigned dec, unsigned width);
 
+/**
+ * @brief
+ * convert an bit vector to its decimal representation
+ * big-endian
+ */
+unsigned bin_to_dec(const std::vector<bool>& bin);
 }  // namespace tket