+++ pytket/docs/changelog.rst
@@ -18,6 +18,10 @@ Minor new features:
   ``Architecture`` classes to give custom name to constructed ``Node``.
 * Add ``FlattenRelabelRegistersPass`` to remove empty quantum wires and relabel all
   qubits to a default register named after a passed label.
+* Updated implementation of ``ToffoliBox`` utilising multiplexors
+  for improved decomposition.
+* Add new ``DiagonalBox`` that synthesises a diagonal unitary matrix
+  into a sequence of multiplexed-Rz gates.
 
 Fixes: